Handle 'peer_storage' msgs and update 'peer_connected'
peer_storage messages (up to 1 KiB of data) from peers with whom we have an active
channel should be stored by our node.

We use PeerState to store peer_storage data for each peer, so we don’t need to manage
forgetting it after all channels with the peer are closed.

Updated 'peer_connected' to send peer_storage, wrapped in 'your_peer_storage',
to the peer if it exists.